Divide common.ts so that Simplenote can use only the label rules
