+++ org/github/milestone.ts
@@ -1,0 +1,9 @@
+import {warn, danger} from "danger";
+
+export default async () => {
+    // Warn if the PR doesn't have a milestone
+    const issue = await danger.github.api.issues.get(danger.github.thisPR);
+    if (issue.data.milestone == null) {
+        warn("PR is not assigned to a milestone.");
+    }
+};
